Create a Settings module for Yaml configuration
While using `ENV.update` works so far, in the future I'll need
non-string key-value pairs (boolean values, for instance), which is not
possible to achieve with ENV because it can only have string-string
pairs.

This commit introduces a module with class methods with names
corresponding to the keys in the settings.yml file.